+++ server/src/main/kotlin/oms/ufsi/dto/CreateFinancialRecordRequest.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class CreateFinancialRecordRequest(
-    val recordType: String, val referenceNumber: String, val amount: Long,
+    val recordType: String, val referenceNumber: String, val amount: Double,
     val currency: String = "EUR", val recordDate: String, val paymentDate: String? = null,
     val description: String? = null, val milestone: String? = null,
     val paymentPurpose: String = "works"
